fix: change model
