+++ domain/model/user.go
@@ -14,7 +14,6 @@ type User struct {
 	bun.BaseModel `bun:"users,alias:u"`
 	ID            string       `bun:"id,pk"`
 	ShopID        string       `bun:"shop_id,notnull"`
-	Shop          *Shop        `bun:"rel:belongs-to,join:shop_id=id"`
 	Name          string       `bun:"name,notnull"`
 	Email         string       `bun:"email,notnull"`
 	Password      string       `bun:"password,notnull"`
@@ -25,20 +24,10 @@ type User struct {
 	DeletedAt     bun.NullTime `bun:"deleted_at,nullzero"`
 }
 
-type Shop struct {
-    bun.BaseModel `bun:"shops,alias:s"`
-    ID            string    `bun:"id,pk"`
-    Name          string    `bun:"name,notnull"`
-    CreatedAt     time.Time `bun:"created_at,notnull"`
-    UpdatedAt     time.Time `bun:"updated_at,notnull"`
-    DeletedAt     time.Time `bun:"deleted_at,nullzero"`
-}
-
 func NewUser(
 	shopID string,
     name string,
 	email string,
-	isShopManager bool,
 ) *User {
 	return &User{
 		ID:            ulid.NewULID(),
@@ -46,37 +35,20 @@ func NewUser(
 		Name:          name,
 		Email:         email,
 		Password:      password.MakeRandomStr(),
-		IsShopManager: isShopManager,
 		CreatedAt:     time.Now(),
 		UpdatedAt:     time.Now(),
 	}
 }
 
-func (u *User) ToDTO() *output.UserDTO {
-	if u == nil {
-        return nil
+func (u *User) ToDTO() *output.User {
+    return &output.User{
+		ID:            u.ID,
+        ShopID:        u.ShopID,
+        Name:          u.Name,
+        Email:         u.Email,
+        Password:      u.Password,
+        LastLoginedAt: ctime.NullTimeToPtrJST(u.LastLoginedAt),
+        CreatedAt:     &u.CreatedAt,
+        UpdatedAt:     &u.UpdatedAt,
     }
-    userDTO := &output.UserDTO{
-        User: output.User{
-			ID:            u.ID,
-            ShopID:        u.ShopID,
-            Name:          u.Name,
-            Email:         u.Email,
-            Password:      u.Password,
-            LastLoginedAt: ctime.NullTimeToPtrJST(u.LastLoginedAt),
-            CreatedAt:     &u.CreatedAt,
-            UpdatedAt:     &u.UpdatedAt,
-        },
-    }
-	if u.Shop != nil {
-        userDTO.Shop = output.Shop{
-            ID:            u.Shop.ID,
-            Name:          u.Shop.Name,
-            CreatedAt:     &u.Shop.CreatedAt,
-            UpdatedAt:     &u.Shop.UpdatedAt,
-        }
-    } else {
-        userDTO.Shop = output.Shop{}
-    }
-    return userDTO
 }
