+++ usecase/input/user.go
@@ -8,5 +8,4 @@ type CreateUserDTO struct {
 	ShopID        string `json:"shopID" validate:"required"`
 	Name          string `json:"name" validate:"required"`
 	Email         string `json:"email" validate:"required"`
-	IsShopManager bool   `json:"isShopManager" validate:"required"`
 }
